+++ react/components/ProductAssemblyOptions/ProductAssemblyOptionItemCustomize.tsx
@@ -2,12 +2,14 @@ import React, { Fragment, useState, FC } from 'react'
 import { Button, Modal } from 'vtex.styleguide'
 import ProductPrice from 'vtex.store-components/ProductPrice'
 import { useDevice } from 'vtex.device-detector'
+import { FormattedMessage } from 'react-intl'
 
 import { useProductAssemblyItem } from '../ProductAssemblyContext/Item'
 import ProductAssemblyOptionsGroup from './ProductAssemblyOptionsGroup'
 import { imageUrlForSize } from './ProductAssemblyOptionItemImage'
 import { ProductAssemblyGroupContextProvider } from '../ProductAssemblyContext/Group'
 import { withItem } from './withItem'
+import StopPropagation from '../StopPropagation'
 
 const IMG_SIZE = 140
 
@@ -102,16 +104,20 @@ const ProductAssemblyOptionItemCustomize: FC<Props> = ({
         collapseLeft={buttonCollapse === 'left'}
         collapseRight={buttonCollapse === 'right'}
       >
-        <div className="c-action-primary t-action">Customize</div>
+        <div className="c-action-primary t-action">
+          <FormattedMessage id="store/product-customizer.customize" />
+        </div>
       </Button>
-      <Modal
-        isOpen={modalOpen}
-        onClose={closeAction}
-        centered={!isMobile}
-        title={`Customize your ${name}`}
-      >
-        <ModalView closeAction={closeAction}>{children}</ModalView>
-      </Modal>
+      <StopPropagation>
+        <Modal
+          isOpen={modalOpen}
+          onClose={closeAction}
+          centered={!isMobile}
+          title={`Customize your ${name}`}
+        >
+          <ModalView closeAction={closeAction}>{children}</ModalView>
+        </Modal>
+      </StopPropagation>
     </Fragment>
   )
 }