+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Hard-coded `customize` string replaced with `FormattedMessage`.
+- Prevent events from modal content being propagated to its react-tree parents.
 
 ## [2.9.4] - 2020-11-30
 ### Fixed